Create two different pages for /me/clients and /clients.
The two pages uses two different endpoints to get and edit the client form, that
now is a separated component.
Moreover, the Token tab is not showed in /me/clients. This inevitably creates
a bug, since clear_access_tokens_on_refresh and require_auth_time will always
be set to false.