Make the ScimUser id field non-optional.
A workaround to remove the id field is necessary when creating a new user because a bug in the backend.
Indeed, if an id is provided during the user creation,
that id is used by the backend to write the user into the database.
Even worse, if an empty (undefined) is sent to the backend, an entry without id is created into the db.